# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Forex Bot API
  version: 0.1.0
  description: |
    REST API for the Forex/Crypto trading bot SaaS.

    **Owner:** Atlas Goro
    **Contract:** This spec is the source of truth. FastAPI runtime spec at `/openapi.json`
    MUST match — drift is enforced in CI.

    **Auth:** Bearer JWT (access token, 15 min TTL). Use `/auth/refresh` to rotate.

    **Errors:** All errors return the `ErrorResponse` shape. See `docs/api/error-contract.md`.
  contact:
    name: Backend (Atlas Goro)
    email: whyman404@gmail.com
  license:
    name: Proprietary
servers:
  - url: https://api.forexbot.app/api/v1
    description: Production
  - url: https://api.staging.forexbot.app/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /broker-accounts:
    post:
      tags: [broker]
      summary: Connect a new broker account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrokerAccountCreateRequest' }
            example:
              broker: exness_mt5
              label: Exness Live Main
              account_type: live
              credentials:
                login: '12345678'
                password: <secret>
                server: Exness-MT5Real
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrokerAccountPublic' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    get:
      tags: [broker]
      summary: List connected broker accounts
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BrokerAccountPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /broker-accounts/{accountId}:
    parameters:
      - { name: accountId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [broker]
      summary: Update broker account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrokerAccountUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrokerAccountPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [broker]
      summary: Delete broker account
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /broker-accounts/{accountId}/test-connection:
    parameters:
      - { name: accountId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [broker]
      summary: Probe broker connectivity using stored credentials
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrokerConnectionTestResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/BadGateway' }
  /strategies:
    get:
      tags: [strategy]
      summary: List strategy catalog
      responses:
        '200':
          description: Catalog
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StrategyPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /strategies/{code}:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    get:
      tags: [strategy]
      summary: Get single strategy by code
      responses:
        '200':
          description: Strategy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /strategy-instances:
    post:
      tags: [strategy-instance]
      summary: Create a strategy instance (draft state)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StrategyInstanceCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    get:
      tags: [strategy-instance]
      summary: List strategy instances of current user
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /strategy-instances/{instanceId}:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [strategy-instance]
      summary: Update strategy instance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StrategyInstanceUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /strategy-instances/{instanceId}/start:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [strategy-instance]
      summary: Transition instance to `running`
      responses:
        '200':
          description: Running
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /strategy-instances/{instanceId}/stop:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [strategy-instance]
      summary: Graceful stop (allow open positions to close per plan)
      responses:
        '200':
          description: Stopping
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /strategy-instances/{instanceId}/kill:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [strategy-instance]
      summary: Emergency stop — flat all positions immediately
      responses:
        '200':
          description: Killed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /strategy-instances/{instanceId}/preflight:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [strategy-instance]
      summary: Run live-trading gate checks without flipping state
      responses:
        '200':
          description: Gate result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GateResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /strategy-instances/{instanceId}/revert-to-paper:
    parameters:
      - { name: instanceId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [strategy-instance]
      summary: Graceful revert from live to paper (closes positions via OMS first)
      responses:
        '200':
          description: Paper
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StrategyInstancePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /users/me/consent:
    post:
      tags: [compliance]
      summary: Record TOS / privacy / marketing consent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsentRequest' }
      responses:
        '201':
          description: Consent recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsentPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [compliance]
      summary: List recorded consents for current user
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ConsentPublic' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /notifications:
    get:
      tags: [notification]
      summary: List notifications for current user
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NotificationPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/{notificationId}/read:
    parameters:
      - { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [notification]
      summary: Mark notification as read
      responses:
        '200':
          description: Marked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        15-minute access token from `/auth/login`. Rotate using `/auth/refresh`.
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PaymentRequired:
      description: Active subscription required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    BadGateway:
      description: Upstream broker error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: AUTH_INVALID_CREDENTIALS
        message:
          type: string
          example: Invalid email or password.
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
          description: Correlate with logs and distributed traces
          example: 5f1f1c2a-3b22-4d04-90b7-d9a4e3a8f2e1
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    BrokerAccountCreateRequest:
      type: object
      required: [broker, label, credentials]
      properties:
        broker: { type: string, enum: [exness_mt5, binance] }
        label: { type: string, minLength: 1, maxLength: 120 }
        account_type: { type: string, enum: [demo, live], default: demo }
        credentials:
          type: object
          additionalProperties: true
          description: Broker-specific. Encrypted at rest (AES-256-GCM envelope).
    BrokerAccountUpdateRequest:
      type: object
      properties:
        label: { type: string, maxLength: 120, nullable: true }
        is_active: { type: boolean, nullable: true }
        credentials: { type: object, additionalProperties: true, nullable: true }
    BrokerAccountPublic:
      type: object
      required: [id, broker, label, account_type, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        broker: { type: string }
        label: { type: string }
        account_type: { type: string }
        is_active: { type: boolean }
        last_connection_check_status: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    BrokerConnectionTestResponse:
      type: object
      required: [ok, broker, account_id]
      properties:
        ok: { type: boolean }
        broker: { type: string }
        account_id: { type: string, format: uuid }
        latency_ms: { type: integer, nullable: true }
        detail: { type: string, nullable: true }
    StrategyPublic:
      type: object
      required: [id, code, name, instrument, timeframe, default_params]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        instrument: { type: string }
        timeframe: { type: string }
        default_params: { type: object, additionalProperties: true }
    StrategyInstanceCreateRequest:
      type: object
      required: [strategy_code, broker_account_id, label]
      properties:
        strategy_code: { type: string }
        broker_account_id: { type: string, format: uuid }
        label: { type: string, minLength: 1, maxLength: 120 }
        params: { type: object, additionalProperties: true }
        risk_config: { type: object, additionalProperties: true }
    StrategyInstanceUpdateRequest:
      type: object
      properties:
        label: { type: string, maxLength: 120, nullable: true }
        params: { type: object, additionalProperties: true, nullable: true }
        risk_config: { type: object, additionalProperties: true, nullable: true }
    StrategyInstancePublic:
      type: object
      required: [id, strategy_id, broker_account_id, label, status, params, risk_config, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        strategy_id: { type: string, format: uuid }
        broker_account_id: { type: string, format: uuid }
        label: { type: string }
        status: { type: string, enum: [draft, running, paused, stopped, killed, errored] }
        params: { type: object, additionalProperties: true }
        risk_config: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    GateCheck:
      type: object
      required: [name, passed]
      properties:
        name: { type: string }
        passed: { type: boolean }
        detail: { type: string, nullable: true }
    GateResult:
      type: object
      required: [passed, failed_checks, checks]
      properties:
        passed: { type: boolean }
        failed_checks: { type: array, items: { $ref: '#/components/schemas/GateCheck' } }
        warnings: { type: array, items: { $ref: '#/components/schemas/GateCheck' } }
        checks: { type: array, items: { $ref: '#/components/schemas/GateCheck' } }
...
    ConsentRequest:
      type: object
      required: [kind, version, accepted]
      properties:
        kind: { type: string, enum: [tos, privacy, marketing, data_processing] }
        version: { type: string }
        accepted: { type: boolean }
    ConsentPublic:
      type: object
      required: [id, kind, version, accepted, created_at]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [tos, privacy, marketing, data_processing] }
        version: { type: string }
        accepted: { type: boolean }
        created_at: { type: string, format: date-time }
    NotificationPublic:
      type: object
      required: [id, kind, title, body, is_read, created_at]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        body: { type: string }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }